Lower FakeParamOp to appropriately sized zeros output.
FakeParamOp is used in CondV2 to signify a value that is defined along one
branch but not the other. These values should only be used along the
corresponding branch. For the lowering to XLA create a zero output of the
appropriate shape and type for the FakeParam (the value should never be used,
so any value could be set, zero is convenient).

This could be extended in future to instead be a dead value so that additional
verification could be done post lowering.

PiperOrigin-RevId: 217229911