Add identity ops as final output, so custom backends like TF-TRT can work with half_plus_two example.
Without this change, TF-TRT can fuse the 'Add's but not the 'Mul's. This is
because the 'Mul's are the outputs of the graph, and if TF-TRT are going to
fuse them then all 'Mul's will be inside the single TRTEngineOp and thus will
have same op name with different output port, which changes the fetches of the
graph. Thus by design it doesn't fuse the outputs.

PiperOrigin-RevId: 218720036