Improve performance of LRNGrad on CPU builds
This commit improves the performance of the CPU version of LRNGrad
by moving a number of computations out of its innermost loop.  The
extent of the resulting performance gains is dependent on the
depth_radius, the compiler and the number of CPUs available.  The
BM_LRNGrad benchmark is showing a 1.47x performance gain for a
depth_radius of 4 and a 1.3x gain for a depth_radius of 2 when
built with gcc 7.3.0 and run on a single core machine.  With
gcc 5.4.0 the results on a single core machine are less impressive
but are nonetheless significant;  1.12x for a depth radius of 4 and
1.10x with a depth radius of 2. With gcc 7.3.0 on a 16 core machine
speed ups of 1.4x and 1.2x can be observed using a depth radius of
4 and 2 respectively.